Update DEBUG_CHANNEL_ID and repository link
Updated the default value of the `DEBUG_CHANNEL_ID` environment variable to a new channel ID. Modified the code to convert this value to an integer if it is a digit. Changed the repository link in the embed field from a GitLab URL to a GitHub URL to reflect the new hosting location.